refactor: 클래스 및 디렉토리 구조 수정 - 컨버터 클래스를 notice 도메인으로 이동 - repository 인터페이스를 도메인 계층으로 이동
