+++ backend/src/main/java/kr/kro/colla/comment/domain/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package kr.kro.colla.comment.repository;
+package kr.kro.colla.comment.domain.repository;
 
-import kr.kro.colla.task.comment.domain.Comment;
+import kr.kro.colla.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
