--- backend/src/main/java/kr/kro/colla/project/project/repository/ProjectRepository.java
@@ -1,7 +1,0 @@
-package kr.kro.colla.project.project.repository;
-
-import kr.kro.colla.project.project.domain.Project;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-}
