+++ backend/src/main/java/kr/kro/colla/task/task/domain/Task.java
@@ -1,6 +1,6 @@
 package kr.kro.colla.task.task.domain;
 
-import kr.kro.colla.task.comment.domain.Comment;
+import kr.kro.colla.comment.domain.Comment;
 import kr.kro.colla.task.history.domain.History;
 import kr.kro.colla.story.domain.Story;
 import kr.kro.colla.task.task_tag.domain.TaskTag;
